Cleanup and function naming
