+++ filter/regexp.go
@@ -68,11 +68,11 @@ func (filter *RegExp) Configure(conf core.PluginConfigReader) error {
 // ApplyFilter check if all Filter wants to reject the message
 func (filter *RegExp) ApplyFilter(msg *core.Message) (core.FilterResult, error) {
 	if filter.expNot != nil && filter.expNot.MatchString(string(msg.Data())) {
-		return filter.MessageRejectResult(), nil
+		return filter.GetFilterResultMessageReject(), nil
 	}
 
 	if filter.exp != nil && !filter.exp.MatchString(string(msg.Data())) {
-		return filter.MessageRejectResult(), nil
+		return filter.GetFilterResultMessageReject(), nil
 	}
 
 	return core.FilterResultMessageAccept, nil
